+++ tests/test_navitem.py
@@ -133,19 +133,34 @@ def test_get_url_override():
     ],
 )
 def test_active(url, req_path, req_params, expected, rf):
-    item = NavItem(title=..., url=url)
+    item = NavItem(title=..., url=f"http://testserver/{url.lstrip('/')}")
 
     req = rf.get(req_path, req_params)
-    print(f"{req.path=}")
 
     assert item.get_active(req) == expected
 
 
-@pytest.mark.parametrize("append_slash", [True, False])
-def test_active_append_slash_setting(append_slash, req):
-    item = NavItem(title=..., url="/test")
+def test_active_different_scheme(rf):
+    item = NavItem(title=..., url="https://testserver/")
 
-    req.path = "/test"
+    req = rf.get("/")
+
+    assert item.get_active(req) is False
+
+
+def test_active_different_domain(rf):
+    item = NavItem(title=..., url="http://different-domain/")
+
+    req = rf.get("/")
+
+    assert item.get_active(req) is False
+
+
+@pytest.mark.parametrize("append_slash", [True, False])
+def test_active_append_slash_setting(append_slash, rf):
+    item = NavItem(title=..., url="http://testserver/test")
+
+    req = rf.get("/test")
 
     with override_settings(APPEND_SLASH=append_slash):
         assert item.get_active(req) is True